Use qualified name for traced functions in Python 3.3+
This is useful when debugging methods with similar names in different classes,
for instance models and forms in Django which both have `clean()` and `save()`
methods which may call each other. If only `clean()` is displayed, one can't
know for sure if it's `Model.clean()` or `Form.clean()`. But in Python 3.3+,
it's really easy to tell them apart by using `func.__qualname__` instead of
`func.__name__`.

The downside is that it also makes names longer for nested functions and
classes (…which is why it's sometimes useful to add `\n` in the regexps in the
unit tests), but IMO it's worth it :)